Sat, Jun 24, 2023, 9:25 PM +01:00

+++ week2/pset2/readability/readability.c
@@ -11,10 +11,10 @@ int main(void)
     int  i = 0;
 
     //NUMBER OF WORDS
-    int numspace = 1;
+    float numspace = 1;
 
     //NUMBER OF SENTENCES
-    int numperiod = 0;
+    float numperiod = 0;
 
 
     do
